requirements.txt
streamlit
pandas
requests
folium
streamlit-folium
google-generativeai
gTTS
scikit-learn
joblib